backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-generativeai>=0.8.0
faiss-cpu>=1.8.0
requests>=2.31.0

